Move command part of /me under commands/
